Move misplaced StandardPagination.tsx
