+++ site/src/app/bills/page.tsx
@@ -4,7 +4,7 @@ import ShortBill from "@/app/components/ShortBill";
 import HumanFriendlyColumn from "@/app/components/HumanFriendlyColumn";
 import StandardStack from "@/app/components/StandardStack";
 import type { Metadata } from "next";
-import StandardPagination from "@/components/StandardPagination";
+import StandardPagination from "@/app/components/StandardPagination";
 import StandardCard from "@/app/components/StandardCard";
 import StandardCardTitle from "@/app/components/StandardCardTitle";
 import StandardCardDescription from "@/app/components/StandardCardDescription";
